deep-researcher/requirements.txt
sentence-transformers==5.1.0
faiss-cpu==1.12.0
numpy==2.3.3
fpdf2==2.7.8
streamlit==1.38.0
